Re-Attatch-Link

+++ E-Learning/js-file/signup.js
@@ -1,16 +1,22 @@
 document.getElementById("login-form").addEventListener("submit", (e) => {
     e.preventDefault();
-    const bookdata = {
-        username: document.getElementById("username").value,
-        email: document.getElementById("email").value,
-        Password: document.getElementById("password").value,
-        select: document.getElementById("select").value,
-    };
 
+    const email = document.getElementById("email").value;
+    const password = document.getElementById("password").value;
     let signdata = JSON.parse(localStorage.getItem('signdata')) || [];
-    signdata.push(bookdata);
-    localStorage.setItem('signdata', JSON.stringify(signdata));
-    alert("Also Login Is Compalsory");
+    const userExists = signdata.some(user => user.email === email);
 
-    window.location.href = '../index-file/login.html';
+    if (userExists) {
+        alert("User already exists! Please log in.");
+    } else {
+        const bookdata = {
+            email: email,
+            password: password 
+        };
+        signdata.push(bookdata);
+        localStorage.setItem('signdata', JSON.stringify(signdata));
+        alert("Signup successful!");
+
+        window.location.href = '../index-file/login.html';
+    }
 });
